+++ index.php
@@ -4,7 +4,7 @@
  *
  * @package     CacheControl
  * @author      dxw
- * @copyright   2020
+ * @copyright   2022
  * @license     MIT
  *
  * @wordpress-plugin
@@ -12,7 +12,8 @@
  * Plugin URI: https://github.com/dxw/dxw-cache-control
  * Description: Set the Cache control headers by content type, taxonomy, template etc.
  * Author: dxw
- * Version: 0.1.0
+ * Version: 0.1.1
+ * Requires at least: 6.1
  * Network: True
  */